pyproject.toml
[tool.poetry]
name = "pi2comercio"
version = "0.1.0"
description = ""
authors = ["moliv <molivramento@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
ormar = {extras = ["postgresql"], version = "^0.12.2"}
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
pytest = "^7.4.3"
httpx = "^0.25.0"
boto3 = "^1.28.77"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
